Day 14: Crash fixed! See description
Crash was caused by an infinite loop in the ShooterController.cs class. It was from the coroutine. Since it was called, I have an infinite loop in there so the code inside of it can run on a constant timer. It never returned anything thus the loop stayed looping forever causing unity to crash.